requirements.txt
pandas>=1.3
openpyxl>=3.0
google-cloud-asset>=2.8.0
google-cloud-storage>=2.7.0
protobuf>=3.20.0
